+++ src/client/src/components/export/exportDialog.tsx
@@ -1,12 +1,15 @@
 import { useTranslation } from "react-i18next";
+import { useSelector } from "react-redux";
 import { Dialog, DialogActions, DialogContent, DialogTitle, Stack, Typography } from "@mui/material";
 import { GridRowSelectionModel } from "@mui/x-data-grid";
+import { ReduxRootState, User } from "../../api-lib/ReduxStateInterfaces.ts";
 import {
   exportCSVBorehole,
   exportGeoPackageBoreholes,
   exportJsonBoreholes,
   exportJsonWithAttachmentsBorehole,
 } from "../../api/borehole.ts";
+import { useAuth } from "../../auth/useBdmsAuth.tsx";
 import { downloadData, downloadDataFromBlob } from "../../utils.ts";
 import { CancelButton, ExportButton } from "../buttons/buttons.tsx";
 
@@ -18,6 +21,9 @@ interface ExportDialogProps {
 }
 export const ExportDialog = ({ isExporting, setIsExporting, selectionModel, fileName }: ExportDialogProps) => {
   const { t } = useTranslation();
+  const auth = useAuth();
+  const user: User = useSelector((state: ReduxRootState) => state.core_user);
+  const canExportAttachments = !auth.anonymousModeEnabled && user.data.roles.includes("EDIT");
 
   const exportJson = async () => {
     const exportJsonResponse = await exportJsonBoreholes(selectionModel);
@@ -53,8 +59,7 @@ export const ExportDialog = ({ isExporting, setIsExporting, selectionModel, file
           <Stack gap={1} sx={{ mt: 3 }}>
             <ExportButton label={"CSV"} onClick={exportCsv} />
             <ExportButton label={"JSON"} onClick={exportJson} />
-            <ExportButton label={"JSON + PDF"} onClick={exportJsonWithAttachments} />
-            <ExportButton label={"ZIP"} onClick={exportJsonWithAttachments} />
+            {canExportAttachments && <ExportButton label={"JSON + PDF"} onClick={exportJsonWithAttachments} />}
             <ExportButton label={"GPKG"} onClick={exportGeoPackage} />
           </Stack>
         </DialogContent>